% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotCo}
\alias{PlotCo}
\title{Plotting crossover positions}
\usage{
PlotCo(
  sample,
  chr = "chr9",
  co,
  scale = "pos",
  export.options = F,
  labels = T,
  label_font = 12,
  label_angle = 0
)
}
\arguments{
\item{sample}{Pollen raw data}

\item{chr}{chromosome name}

\item{co}{crossover table}

\item{scale}{data locus or real position}

\item{labels}{annotate label or not}

\item{label_font}{label font}

\item{label_angle}{label position}
}
\value{
crossover plot
}
\description{
Plotting crossover positions
}
\examples{
library(IIIandme)
input<- PreProcessing(sample)
res<- HapCo(input, 5)
vote<- VoteCount(sample, input, res, 5)
PlotCo(sample, chr='chr9', co=vote[[2]])
}
