% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFuns.R
\name{PreProcessing}
\alias{PreProcessing}
\title{Preprocessing & Basic functions}
\usage{
PreProcessing(data)
}
\arguments{
\item{data}{raw data as the sample data}
}
\value{
output of preprocessing data
}
\description{
Preprocessing & Basic functions
}
\examples{
library(IIIandMe)
PreProcessing(sample)
}
