% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteHap.R
\name{VoteCount}
\alias{VoteCount}
\title{Votting the final haplotype}
\usage{
VoteCount(sample, input, hap_co, npol)
}
\arguments{
\item{input}{Pollen data after pre-processing}

\item{hap_co}{haplotype & co lists}

\item{npol}{pollen numbers}
}
\value{
haplotype lists
}
\description{
Votting the final haplotype
}
\examples{
library(IIIandMe)
input<- PreProcessing(sample)
res<- HapCo(input, 5)
hap<- res[[1]]
co<- res[[2]]
VoteCount(sample, input, res, 5)
}
