% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phasing.R
\name{HapCo}
\alias{HapCo}
\title{Generating all possible haplotypes & positions of crossovers}
\usage{
HapCo(data, npol, filterGenoError = F)
}
\arguments{
\item{data}{preprocessing data}

\item{npol}{pollen numbers}

\item{filterGenoError}{filtering genotyping errors}
}
\value{
lists of possible haplotypes & crossovers
}
\description{
Generating all possible haplotypes & positions of crossovers
}
\examples{
library(IIIandme)
input<- PreProcessing(sample)
HapCo(input, 5)
}
