% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteHap.R
\name{VoteCount}
\alias{VoteCount}
\title{Votting the final haplotype}
\usage{
VoteCount(hap, npol, name)
}
\arguments{
\item{hap}{all possible haplotypes}

\item{npol}{pollen numbers}

\item{name}{names of pollens}
}
\value{
haplotype lists
}
\description{
Votting the final haplotype
}
\examples{
library(IIIandme)
input<- PreProcessing(sample)
res<- HapCo(input, 5)
hap<- res[[1]]
co<- res[[2]]
VoteCount(hap, 5, colnames(input))
}
